User commit

+++ src/main/java/com/example/demo/service/Impl/ValidationServiceImpl.java
@@ -17,7 +17,7 @@ public ValidationEntity postData(ValidationEntity stu){
     return student.save(stu);
 }
 @Override
-public ValidationEntity getStudentById(Integer id){
+public ValidationEntity getStudentById(Long id){
     return student.findById(id).orElseThrow(()->new ValidationException("Invalid Id"+id));
 }
 }
